Move `testdata_test.go` to new package `testutil/testdata`
A fair amount of package juggling was necessary to avoid circular imports:

- Methods on `scan.Dir` and `scan.File` had to be exported because the simulation methods of `Node` uses them.
- The file `scan/run_test.go` had to use package `scan_test` to use `testdata` (and the other new package `scantest`) which uses the types in `scan`.
- The assertion functions were moved from `assert_test.go` to another new package `scan/scantest` because they had to be made public due to their use in `testdata`: It didn't feel right to have test-specific functions/methods exposed directly from `scan`.
